Move handler under ee folder
